Fix a few small bugs and re-work tags to be cleaner
