+++ EXIF Rewrite/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="2.9.8" targetFramework="net40" developmentDependency="true" />
   <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.8" targetFramework="net40" developmentDependency="true" />
   <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.8" targetFramework="net40" developmentDependency="true" />
   <package id="Microsoft.CodeQuality.Analyzers" version="2.9.8" targetFramework="net40" developmentDependency="true" />
